todos in the it specs
