+++ it/endpoints/CheckWithNinoISpec.scala
@@ -1,0 +1,42 @@
+package endpoints
+
+import play.api.http.Status.OK
+import stubs.HomeOfficeImmigrationStatusStubs
+import support.ISpec
+
+class CheckWithNinoISpec extends ISpec with HomeOfficeImmigrationStatusStubs {
+
+  "GET /check-immigration-status/check-with-nino" should {
+    "show the lookup page" in {
+      givenAuthorisedForStride("TBC", "StrideUserId")
+
+      val result = request("/check-with-nino").get().futureValue
+
+      result.status shouldBe OK
+      result.body should include(htmlEscapedMessage("lookup.title"))
+      result.headers.get("Cache-Control").map(_.mkString) shouldBe Some("no-cache, no-store, must-revalidate")
+    }
+  }
+
+  "POST /check-immigration-status/check-with-nino" should {
+    "redirect to the result page" in {
+      givenStatusCheckSucceeds()
+      givenAuthorisedForStride("TBC", "StrideUserId")
+
+      val payload = Map(
+        "dateOfBirth.year"  -> "2001",
+        "dateOfBirth.month" -> "01",
+        "dateOfBirth.day"   -> "31",
+        "familyName"        -> "Jane",
+        "givenName"         -> "Doe",
+        "nino"              -> nino.nino)
+
+      val sessionId = "123"
+      val result = request("/check-with-nino", sessionId).post(payload).futureValue
+
+      result.status shouldBe OK
+      result.body should include(htmlEscapedMessage("status-found.title"))
+      result.headers.get("Cache-Control").map(_.mkString) shouldBe Some("no-cache, no-store, must-revalidate")
+    }
+  }
+}
