--- it/support/AppISpec.scala
@@ -1,5 +1,0 @@
-package support
-
-import org.scalatestplus.play.guice.GuiceOneAppPerSuite
-
-abstract class AppISpec extends BaseISpec with GuiceOneAppPerSuite
