--- it/support/ServerISpec.scala
@@ -1,5 +1,0 @@
-package support
-
-import org.scalatestplus.play.guice.GuiceOneServerPerSuite
-
-abstract class ServerISpec extends BaseISpec with GuiceOneServerPerSuite
